% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fid_data_wrangler.R
\name{getProjectionValue}
\alias{getProjectionValue}
\title{This function calculate projection value:}
\usage{
getProjectionValue(DF)
}
\arguments{
\item{DF}{with}

\item{Date}{and}

\item{Closed}{price and returns a}
}
\value{
projection value of the fund from the start date and assuming 10k investment from the start and reinvest all earnings
}
\description{
This function calculate projection value:
}
