% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fid_goodness_of_fit.R
\name{md_corr_test}
\alias{md_corr_test}
\title{correlation test, pearson or spearman}
\usage{
md_corr_test(md, method = "pearson")
}
\arguments{
\item{md}{lm model}

\item{method}{correlation test method, pearson or spearman}
}
\description{
correlation test, pearson or spearman
}
