% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fid_goodness_of_fit.R
\name{md_f_test}
\alias{md_f_test}
\title{F-test for two sample standard errors}
\usage{
md_f_test(md)
}
\arguments{
\item{md}{lm model}
}
\value{
variance F test
}
\description{
F-test for two sample standard errors
}
