% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fid_goodness_of_fit.R
\name{md_ks_test}
\alias{md_ks_test}
\title{test y_sim and actual y using ks test,low p-value will reject the hypo-thesis that they are from the same distribution}
\usage{
md_ks_test(md)
}
\arguments{
\item{md}{lm model}
}
\value{
ks test between y_obs and y_sim
}
\description{
test y_sim and actual y using ks test,low p-value will reject the hypo-thesis that they are from the same distribution
}
