% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fid_goodness_of_fit.R
\name{md_perm_test}
\alias{md_perm_test}
\title{test y_sim and actual y using permutation test, low p-value will reject the hypothesis that they are independent}
\usage{
md_perm_test(md)
}
\arguments{
\item{md}{lm model}
}
\value{
permutation results
}
\description{
test y_sim and actual y using permutation test, low p-value will reject the hypothesis that they are independent
}
