% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fid_goodness_of_fit.R
\name{md_t_test}
\alias{md_t_test}
\title{test y_sim and actual y using t-statistics, low p-value reject the hypothesis that they have same mean.}
\usage{
md_t_test(md)
}
\arguments{
\item{md}{lm model}
}
\value{
t test result
}
\description{
test y_sim and actual y using t-statistics, low p-value reject the hypothesis that they have same mean.
}
