% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fid_goodness_of_fit.R
\name{md_test}
\alias{md_test}
\title{simulate and test bundle}
\usage{
md_test(md, method = "pearson", modelequation = md$call)
}
\arguments{
\item{md}{lm model}

\item{method}{correlation test's method, pearson, spearman, etc.}
}
\value{
between y_obs and y_sim: ks.test, permutation test, t.test, correlation test, variance F-test and histogram.
}
\description{
Use the above function to simulate y_sim and compare to the original y_obs
}
