% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fid_goodness_of_fit.R
\name{md_y_sim}
\alias{md_y_sim}
\title{Simulate y}
\usage{
md_y_sim(md)
}
\arguments{
\item{md}{lm model}
}
\value{
simulated y_sim
}
\description{
using predicted y value as mean and calculated sd as sd simulate y
}
