% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fid_goodness_of_fit.R
\name{plot_md_hist}
\alias{plot_md_hist}
\title{plot the observed data and simulated data as histogram on the same graph}
\usage{
plot_md_hist(md, modelequation = md$call)
}
\arguments{
\item{md}{lm model}

\item{modelequation}{model description for plotting purpose}
}
\value{
a histogram of y_obs and y_sim
}
\description{
plot the observed data and simulated data as histogram on the same graph
}
