% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fid_goodness_of_fit.R
\name{plotmodel}
\alias{plotmodel}
\title{plot model}
\usage{
plotmodel(md, modelequation = md$call, linecolor = "#CE8891")
}
\arguments{
\item{md:}{a lm model,}

\item{modelequation:}{a string that describe the model(if not specified, will be the $call of the lm model)}

\item{linecolor:}{specify a color for the line occurs in the plot
The two plots are in list, and it is a ggplot so
one can modify the plot by adding theme to the output plots}
}
\value{
Standardized residual plot and residual QQ plot
}
\description{
plot model
}
