% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fid_data_wrangler.R
\name{standardizedNAV}
\alias{standardizedNAV}
\title{function for standarize NAV
takes:}
\usage{
standardizedNAV(DF)
}
\arguments{
\item{DF}{dataframe with Dates and Close price}
}
\value{
Standarized Close price calculated into z-score
}
\description{
function for standarize NAV
takes:
}
