% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fid_goodness_of_fit.R
\name{md_std}
\alias{md_std}
\title{generate SD for each observed datapoint}
\usage{
md_std(md)
}
\arguments{
\item{md}{a lm model with 1 predictor and no intercept}
}
\value{
adjusted standard error for each X in the original model
}
\description{
generate SD for each observed datapoint
}
