% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fid_data_wrangler.R
\name{sqerr}
\alias{sqerr}
\title{Calculate squared distance between two lines}
\usage{
sqerr(x, y)
}
\arguments{
\item{x}{a line}

\item{y}{another line}
}
\value{
Calculate Euclidean distances between two sets of data
}
\description{
Calculate squared distance between two lines
}
